Add back arrow to App Settings modal, move Nutrition Goals to dedicated modal, fix WeightChart/WeightInsights card widths
