+++ app/_layout.tsx
@@ -16,7 +16,7 @@ function RootNavigator() {
     if (isLoading) return;
 
     const onboardingComplete = preferences.onboardingComplete === true;
-    const inApp = segments[0] === '(tabs)' || segments[0] === 'add-food-modal' || segments[0] === 'create-meal-modal' || segments[0] === 'app-settings-modal' || segments[0] === 'appearance-modal';
+    const inApp = segments[0] === '(tabs)' || segments[0] === 'add-food-modal' || segments[0] === 'create-meal-modal' || segments[0] === 'app-settings-modal' || segments[0] === 'appearance-modal' || segments[0] === 'nutrition-goals-modal';
 
     if (onboardingComplete && !inApp) {
       const tab = preferences.defaultTab ?? 'nutrition';
@@ -75,6 +75,13 @@ function RootNavigator() {
       />
       <Stack.Screen
         name="appearance-modal"
+        options={{
+          presentation: 'modal',
+          headerShown: false,
+        }}
+      />
+      <Stack.Screen
+        name="nutrition-goals-modal"
         options={{
           presentation: 'modal',
           headerShown: false,
